+++ src/old/Paladins.ts
@@ -1,0 +1,3 @@
+import { API } from './API';
+
+export default API;
