+++ src/paladins.spec.ts
@@ -1,0 +1,5 @@
+import API from './paladins';
+
+test('should 1 equal 1', () => {
+    expect(1).toBe(1);
+});
